requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-jose
passlib[bcrypt]
pydantic
alembic
